backend/requirements.local.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
geopy==2.4.1
reportlab==4.0.8
django-cors-headers==4.3.1
requests==2.31.0